+++ src/pymovements/events/__init__.py
@@ -29,9 +29,6 @@ from pymovements.events.detection.library import register_event_detection
 from pymovements.events.events import Events
 from pymovements.events.frame import EventDataFrame
 from pymovements.events.precomputed import PrecomputedEventDataFrame
-from pymovements.events.segmentation import events2segmentation
-from pymovements.events.segmentation import events2timeratio
-from pymovements.events.segmentation import segmentation2events
 
 
 __all__ = [
@@ -43,9 +40,6 @@ __all__ = [
     'ivt',
     'microsaccades',
     'out_of_screen',
-    'events2segmentation',
-    'events2timeratio',
-    'segmentation2events',
 
     'PrecomputedEventDataFrame',
     'Events',
