+++ tests/unit/transforms/norm_test.py
@@ -59,7 +59,7 @@ import pymovements as pm
 )
 def test_norm_of_two_columns_returns(columns, df, expected_series):
     result_df = df.select(
-        pm.gaze.transforms.norm(columns=columns).alias('norm'),
+        pm.transforms.norm(columns=columns).alias('norm'),
     )
     assert_series_equal(result_df['norm'], expected_series, check_names=False)
 
@@ -72,7 +72,7 @@ def test_norm_of_list_column_returns():
     )
 
     result_df = df.select(
-        pm.gaze.transforms.norm(
+        pm.transforms.norm(
             column='velocity',
             components=(0, 1),
         ).alias('norm'),
@@ -93,7 +93,7 @@ def test_norm_of_struct_column_returns():
     )
 
     result_df = df.select(
-        pm.gaze.transforms.norm(
+        pm.transforms.norm(
             column='velocity',
             components=('x', 'y'),
         ).alias('norm'),
@@ -114,7 +114,7 @@ def test_norm_raises_for_missing_parent_column():
 
     with pytest.raises(pl.exceptions.ColumnNotFoundError, match='position'):
         df.select(
-            pm.gaze.transforms.norm(
+            pm.transforms.norm(
                 column='position',
                 components=('x', 'y'),
             ).alias('norm'),
@@ -132,7 +132,7 @@ def test_norm_raises_for_missing_field_in_parent_column():
 
     with pytest.raises(pl.exceptions.StructFieldNotFoundError, match='z'):
         df.select(
-            pm.gaze.transforms.norm(
+            pm.transforms.norm(
                 column='velocity',
                 components=('x', 'z'),
             ).alias('norm'),
@@ -148,7 +148,7 @@ def test_norm_raises_when_parent_column_is_not_struct():
 
     with pytest.raises(pl.exceptions.PolarsError):
         df.select(
-            pm.gaze.transforms.norm(
+            pm.transforms.norm(
                 column='velocity',
                 components=('x', 'y'),
             ).alias('norm'),
@@ -158,7 +158,7 @@ def test_norm_raises_when_parent_column_is_not_struct():
 def test_norm_raises_when_components_not_length_2():
     message = 'components must be of length 2 but is 3'
     with pytest.raises(ValueError, match=message):
-        pm.gaze.transforms.norm(column='velocity', components=('x', 'y', 'z'))
+        pm.transforms.norm(column='velocity', components=('x', 'y', 'z'))
 
 
 @pytest.mark.parametrize(
@@ -171,16 +171,16 @@ def test_norm_raises_when_components_not_length_2():
 def test_norm_raises_when_component_elements_unexpected_type(components):
     message = "elements of 'components' must be either of type int or str but they are"
     with pytest.raises(TypeError, match=message):
-        pm.gaze.transforms.norm(column='velocity', components=components)
+        pm.transforms.norm(column='velocity', components=components)
 
 
 def test_norm_raises_when_components_unexpected_type():
     message = "'components' must be a sequence but is of type float"
     with pytest.raises(TypeError, match=message):
-        pm.gaze.transforms.norm(column='velocity', components=0.1)
+        pm.transforms.norm(column='velocity', components=0.1)
 
 
 def test_norm_raises_when_column_and_columns_are_none():
     message = 'either column or columns must be provided but both are None'
     with pytest.raises(TypeError, match=message):
-        pm.gaze.transforms.norm(column=None, columns=None)
+        pm.transforms.norm(column=None, columns=None)
