+++ tests/unit/transforms/resample_test.py
@@ -658,7 +658,7 @@ import pymovements as pm
 )
 def test_resample_returns(kwargs, df, expected_df):
     """Test if resample returns expected DataFrame."""
-    result_df = pm.gaze.transforms.resample(df, **kwargs)
+    result_df = pm.transforms.resample(df, **kwargs)
 
     assert_frame_equal(result_df, expected_df)
 
@@ -724,7 +724,7 @@ def test_resample_raises_error(kwargs, exception, msg_substrings):
     })
 
     with pytest.raises(exception) as excinfo:
-        pm.gaze.transforms.resample(df, **kwargs)
+        pm.transforms.resample(df, **kwargs)
 
     msg, = excinfo.value.args
     for msg_substring in msg_substrings:
