+++ tests/unit/transforms/segmentation_test.py
@@ -25,10 +25,10 @@ import polars as pl
 import pytest
 from polars.testing import assert_frame_equal
 
-from pymovements.events.segmentation import _has_overlap
-from pymovements.events.segmentation import events2segmentation
-from pymovements.events.segmentation import events2timeratio
-from pymovements.events.segmentation import segmentation2events
+from pymovements.transforms.segmentation import _has_overlap
+from pymovements.transforms.segmentation import events2segmentation
+from pymovements.transforms.segmentation import events2timeratio
+from pymovements.transforms.segmentation import segmentation2events
 
 
 @pytest.fixture(name='events_df')
